requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
groq
python-jose
passlib
python-multipart 